Mag reading to double, ekf reads opposite of x axis for heading??

+++ RocketFlightComputer/src/utility.hpp
@@ -45,9 +45,9 @@ public:
         float gyroX;
         float gyroY;
         float gyroZ;
-        float magX;
-        float magY;
-        float magZ;
+        double magX;
+        double magY;
+        double magZ;
         float pressure;
 
         // Calculated Values
